decopule YamlServiceProcessor
