+++ packages/autodiscovery/src/Yaml/ExplicitToAutodiscoveryConverter.php
@@ -7,7 +7,6 @@ namespace Symplify\Autodiscovery\Yaml;
 use Nette\Utils\Strings;
 use ReflectionClass;
 use Symfony\Component\Filesystem\Filesystem;
-use Symplify\Autodiscovery\Arrays;
 use Symplify\Autodiscovery\Exception\ClassLocationNotFoundException;
 use Symplify\Autodiscovery\Exception\ClassNotFoundException;
 use Symplify\SmartFileSystem\SmartFileInfo;
@@ -18,11 +17,6 @@ final class ExplicitToAutodiscoveryConverter
      * @var bool
      */
     private $enableAutowire = false;
-
-    /**
-     * @var bool
-     */
-    private $removeService = false;
 
     /**
      * @var bool
@@ -57,18 +51,18 @@ final class ExplicitToAutodiscoveryConverter
     private $commonNamespaceResolver;
 
     /**
-     * @var TagAnalyzer
+     * @var YamlServiceProcessor
      */
-    private $tagAnalyzer;
+    private $yamlServiceProcessor;
 
     public function __construct(
         Filesystem $filesystem,
         CommonNamespaceResolver $commonNamespaceResolver,
-        TagAnalyzer $tagAnalyzer
+        YamlServiceProcessor $yamlServiceProcessor
     ) {
         $this->filesystem = $filesystem;
         $this->commonNamespaceResolver = $commonNamespaceResolver;
-        $this->tagAnalyzer = $tagAnalyzer;
+        $this->yamlServiceProcessor = $yamlServiceProcessor;
     }
 
     /**
@@ -85,7 +79,7 @@ final class ExplicitToAutodiscoveryConverter
         }
 
         foreach ($yaml[YamlKey::SERVICES] as $name => $service) {
-            $yaml = $this->processService($yaml, $service, $name, $filter);
+            $yaml = $this->yamlServiceProcessor->process($yaml, $service, $name, $filter);
         }
 
         $yaml = $this->completeAutodiscovery($yaml, $filePath, $nestingLevel);
@@ -106,39 +100,6 @@ final class ExplicitToAutodiscoveryConverter
         $this->classes = [];
         $this->enableAutowire = false;
         $this->enableAutoconfigure = false;
-    }
-
-    /**
-     * @param mixed[] $yaml
-     * @param string|mixed[]|null $service
-     * @return mixed[]
-     */
-    private function processService(array $yaml, $service, string $name, string $filter): array
-    {
-        $this->removeService = false;
-
-        if ($this->shouldSkipService($service, $name, $filter)) {
-            return $yaml;
-        }
-
-        if (is_array($service)) {
-            [$yaml, $service, $name] = $this->processArrayService($yaml, $service, $name);
-        }
-
-        // anonymous service
-        if ($service === null) {
-            $this->classes[] = $name;
-            $this->removeService = true;
-        }
-
-        // update
-        if ($this->removeService) {
-            unset($yaml[YamlKey::SERVICES][$name]);
-        } else {
-            $yaml[YamlKey::SERVICES][$name] = $service;
-        }
-
-        return $yaml;
     }
 
     /**
@@ -203,71 +164,6 @@ final class ExplicitToAutodiscoveryConverter
     }
 
     /**
-     * @param mixed|mixed[] $service
-     */
-    private function shouldSkipService($service, string $name, string $filter): bool
-    {
-        $class = $service['class'] ?? $name;
-
-        // skip no-namespace class naming
-        if (! Strings::contains($class, '\\')) {
-            return true;
-        }
-
-        if ($filter && ! Strings::contains($class, $filter)) {
-            return true;
-        }
-
-        // is in vendor?
-        if (class_exists($class)) {
-            $reflectionClass = new ReflectionClass($class);
-            if (Strings::match($reflectionClass->getFileName(), '#/vendor/#')) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * @param mixed[] $yaml
-     * @param mixed[] $service
-     * @return mixed[]
-     */
-    private function processArrayService(array $yaml, array $service, string $name): array
-    {
-        $service = $this->processAutowire($service);
-        $service = $this->processTags($service);
-
-        // is only named services
-        if (Arrays::hasOnlyKey($service, 'class')) {
-            unset($yaml[YamlKey::SERVICES][$name]);
-
-            $name = $service['class'];
-            $service = null;
-            $yaml[YamlKey::SERVICES][$name] = $service;
-        }
-
-        // is named service
-        if (isset($service['class']) && is_string($name) && ! ctype_upper($name[0]) && ! class_exists($name)) {
-            // @todo check is no where used in the script, regular would do
-
-            unset($yaml[YamlKey::SERVICES][$name]);
-            $name = $service['class'];
-            unset($service['class']);
-
-            $yaml[YamlKey::SERVICES][$name] = $service;
-        }
-
-        // normalize empty service
-        if ($service === []) {
-            $service = null;
-        }
-
-        return [$yaml, $service, $name];
-    }
-
-    /**
      * @param string[] $services
      * @param string[] $commonNamespaces
      * @return string[][]
@@ -321,39 +217,6 @@ final class ExplicitToAutodiscoveryConverter
         sort($excludedDirectories);
 
         return $excludedDirectories;
-    }
-
-    /**
-     * @param mixed[] $service
-     * @return mixed[]
-     */
-    private function processAutowire(array $service): array
-    {
-        // remove autowire
-        if (isset($service[YamlKey::AUTOWIRE])) {
-            unset($service[YamlKey::AUTOWIRE]);
-            $this->enableAutowire = true;
-        }
-
-        return $service;
-    }
-
-    /**
-     * @param mixed[] $service
-     * @return mixed[]
-     */
-    private function processTags(array $service): array
-    {
-        if (! isset($service[YamlKey::TAGS])) {
-            return $service;
-        }
-
-        if ($this->tagAnalyzer->isAutoconfiguredTags($service[YamlKey::TAGS])) {
-            unset($service[YamlKey::TAGS]);
-            $this->enableAutoconfigure = true;
-        }
-
-        return $service;
     }
 
     private function getRelativeClassLocation(string $class, string $configFilePath): string
